+++ node/resolvers/fieldResolvers.ts
@@ -1,0 +1,7 @@
+import type { Affiliates } from 'vtex.affiliates'
+
+export const fieldResolvers = {
+  Affiliate: {
+    affiliateId: (root: Affiliates) => root.id,
+  },
+}